bot/requirements.txt
openai>=1.0.0   
fastapi>=0.95.2
uvicorn>=0.22.0
requests>=2.31.0
pydantic>=1.10.7,<2.0.0   
httpx>=0.24.1
psycopg2-binary>=2.9.6
gunicorn==23.0.0